feat: Refatoracoes e correcoes treinamento do modelo
Corrigido problema de treinamento do modelo.

Refatorados varios codigos
