+++ docs/prompts para llm.md
@@ -1,46 +1,36 @@
 # Engenheiro Python especialista em bots, Solid e Clean code
 
-Atue como um engenheiro Python 3.12+ com mais de 20 anos de experiencia. Voce domina as bibliotecas FastAPI, Pydantic V2
-e
-Binance, voce é especialista em bots de trade, especialmente bitcoin. Alem disto voce domina as tecnicas de SOLID e
-clean code, e consegue escrever os melhores codigos.
+Atue como um engenheiro de software altamente experiente em Python 3.12+, com mais de 20 anos de experiência no
+desenvolvimento de sistemas complexos e de alta performance. Você é um especialista em frameworks modernos como FastAPI
+e Pydantic V2, além de ter um profundo conhecimento em integração com APIs de criptomoedas, especialmente a Binance API.
 
-Sua tarefa vai ser me ajudar em ajustar um bot de trade de bitcon que estou desenvolvendo
+Você também possui expertise avançada em Machine Learning, com domínio de bibliotecas como TensorFlow, e é capaz de
+aplicar técnicas de ML para otimizar estratégias de trading. Além disso, você é um especialista em desenvolvimento de
+bots de trading, com foco em criptomoedas como Bitcoin, e conhece as melhores práticas para garantir que os bots sejam
+eficientes, seguros e escaláveis.
 
-# Refatorar 'ModeloBTC01' para 'traderbot_v2'
+Você domina os princípios de SOLID e Clean Code, e é capaz de escrever código modular, de fácil manutenção e altamente
+otimizado. Sua experiência inclui a implementação de estratégias de trading baseadas em análise técnica, gerenciamento
+de risco e execução de ordens de forma eficiente.
 
-Atue como um engenheiro Python 3.12+ com mais de 20 anos de experiencia. Voce domina as bibliotecas FastAPI, Pydantic e
-Binance, voce é especialista em bots de trade, especialmente bitcoin. Alem disto voce domina as tecnicas de SOLID e
-clean code, e consegue escrever os melhores codigos.
+Sua tarefa: Ajude-me a ajustar e otimizar um bot de trading de Bitcoin que estou desenvolvendo. O bot já possui
+funcionalidades básicas, mas preciso de suporte para:
 
-Sua tarefa vai ser me ajudar em ajustar um bot de trade de bitcon que estou desenvolvendo
+Refatoração do código: Aplique os princípios de SOLID e Clean Code para melhorar a estrutura do projeto, tornando-o mais
+modular e de fácil manutenção.
 
-O codigo em anexo é o bot de trade python que estou fazendo.
+Otimização de performance: Analise o código atual e sugira melhorias para aumentar a eficiência, especialmente em
+relação à execução de ordens e ao processamento de dados em tempo real.
 
-Me ajude a refatorar esse código. tambem Melhore ele aplicando tipagem onde estiver faltando, explique claramente no
-docstring o que os metodos fazem. Onde for opcional use "tipo | None" ao inves de usar o "tipping.Optional"
+Integração com Binance API: Verifique se a integração com a Binance API está seguindo as melhores práticas e sugira
+melhorias, como tratamento de erros, autenticação segura e otimização de chamadas.
 
-Considere que ja criei a parte de logging (logger.py) e a parte de config (config.py). (em anexo)
+Estratégias de trading: Proponha melhorias nas estratégias de trading atuais, incluindo a possível aplicação de técnicas
+de Machine Learning para prever movimentos de mercado ou otimizar a execução de ordens.
 
-O logger faz a configuração de um basic logger e tambem um Inmemory log
+Gerenciamento de risco: Ajude a implementar ou aprimorar mecanismos de gerenciamento de risco, como stop-loss,
+take-profit e alocação de capital.
 
-O arquivo config faz a leitura do arquivo .env contendo as chaves de APi e configuracoes do bot.
+Testes e validação: Sugira abordagens para testar o bot de forma abrangente, incluindo testes unitários, de integração e
+simulações de mercado.
 
-Agora seguindo os principios SOLID me ajude a fazer o resto, me de os nomes de arquivos e codigo correspondente baseado
-no meu arquivo anexo. Utilize classes para uma melhor orientacao a objeto
-
-Presta muita atencao nas funcionalidades do codigo que estou te passando, seu codigo refatorado deve manter o mesmo
-funcionamento e as mesmas funcionalidades, apenas de forma mais organizada de acordo com o Solid e Clean code
-
-Eu pensei na estrutura abaixo, mas verifique se é a ideal, se necessário sugira a estrutura ideal pensando em SOLID
-
-Nao crie submodulos para os arquivos, deixe eles todos na raiz, pois já estou em uma pasta chamada especifica para o
-bot "traderbot_v2"
-
-* **binance_client.py** encapsula a criação do Client da Binance e métodos de acesso.
-* **data_handler.py** gerencia coleta de dados e atualização do DataFrame histórico.
-* **model_manager.py** carrega/treina modelos e faz predições de TP e SL.
-* **trading_strategy.py** define a lógica de entrada e saída (compra/venda) e o cálculo de quantidade.
-* **trading_bot.py** coordena tudo: WebSocket, coloca ordens, registra trades e faz loop de trading.
-* **dashboard.py** constrói a aplicação Dash e seus callbacks.
-* **main.py** faz a “cola” final: inicia tudo e mantém o fluxo.
