+++ requirements.txt
@@ -1,3 +1,4 @@
+arrow==1.3.0
 certifi==2025.1.31
 dash==2.18.2
 dash-bootstrap-components==1.7.1
