+++ src/core/constants.py
@@ -19,3 +19,7 @@ FEATURE_COLUMNS = [
 # - / "trained_models": Adiciona a pasta "trained_models" ao caminho.
 TRAINED_MODELS_DIR = Path(__file__).resolve().parent.parent / "trained_models"
 TRAINED_MODELS_DIR.mkdir(parents=True, exist_ok=True)
+
+# Define o caminho absoluto para a pasta de dados de treino
+TRAIN_DATA_DIR = Path(__file__).resolve().parent.parent / "train_data"
+TRAIN_DATA_DIR.mkdir(parents=True, exist_ok=True)
